cleanup and refactor into event dir
