+++ program-libs/compressed-account/src/instruction_data/insert_into_queues.rs
@@ -59,15 +59,6 @@ pub struct MerkleTreeSequenceNumber {
     pub seq: U64,
 }
 
-impl From<&MerkleTreeSequenceNumber> for crate::event::MerkleTreeSequenceNumber {
-    fn from(seq: &MerkleTreeSequenceNumber) -> Self {
-        Self {
-            pubkey: seq.pubkey.into(),
-            seq: seq.seq.into(),
-        }
-    }
-}
-
 #[derive(Debug, Clone, PartialEq)]
 pub struct InsertIntoQueuesInstructionData<'a> {
     meta: Ref<&'a [u8], InsertIntoQueuesInstructionDataMeta>,
